optee-client_%.bbappend: add teeclnt group, rpmb_cid and tpm_ftpm_tee module unload
This is a squash of a couple of various fixes from Jerome. See verbose
commits below.

optee-client: add teeclnt group for TEE client applications

Due /dev/tee0 permissions, TEE applications currently have to run as root.
Introduce the teeclnt group so that application do not need full root
priviledges to access the TEE.
- the group is created by optee-client_%.bbappend
- ledge-minimal.bb is updated to add user 'ledge' to the 'teeclnt' group
for convenience (when OP-TEE is enabled)
- Permissions on device /dev/tee0 are set via udev rules, added to
optee-client_%.bbappend

Note that tee-supplicant is still run as root for simplicity and because
it seems there is not much security to gain from running it as a normal
user. It would still need CAP_SYS_RAWIO to access RPMB which is quite a
powerful capability already, and running as non-root introduces some
complexity (notably related to setting permissions on /dev/teepriv0 and
the startup of tee-supplicant, i.e., a dependency between a udev rule and
the tee-supplicant systemd service).

optee-client_%.bbappend: avoid variable expansion in sed expression

$RPMB_CID must not be expanded when the sed expression is executed. Use
single quotes instead of double quotes to achieve that goal.

optee-client: systemd: kill initramfs tee-supplicant before starting new instance

The instance of tee-supplicant started by the initramfs (init.cryptfs) is
not suitable for running normal Trusted Applications because it cannot
access the root FS where .ta files are stored. Therefore, make sure that
this temporary instance is killed just before systemd launches the
tee-supplicant service.

optee-client_%.bbappend: unload tpm_ftpm_tee module before stopping tee-supplicant

If the tpm_ftpm_tee module is unloaded from the kernel when tee-supplicant
is not running, the fTPM TA fails to complete some command and there is a
TA panic dump on the console. This happens for example on system shutdown
or reboot. To avoid this, make sure that the module is unloaded before
stopping tee-supplicant. Tested on RockPi 4.

https://github.com/Linaro/meta-ledge/commit/1cc240312b782198bc980a3ba6c7b766e8d2693c
https://github.com/Linaro/meta-ledge/commit/e774c29979f93d9fd677499b6e3d054233e049eb
https://github.com/Linaro/meta-ledge/commit/b0f5903496e51701549bc0c6230598ceb7ed112f

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Peter Griffin <peter.griffin@linaro.org>