linux-ledge-common.inc: update to use linux-yocto
Updates to use linux-yocto kernel which is used by ewaol.
As part of this we convert to use kmeta cfg fragments like
ewaol and linux-yocto are doing to enable kernel features.

However the last defconfig kmeta is a copy of aarch64 defconfig
so as to enable all Arm platforms and have a multi platform kernel.
We may want to revisit that approach, but without it you end up
with a kernel that only support VEXPRESS.

Other BSPs like xilinux/renesas  for example would add a rockchip.cfg to
linux-yocto kmeta repo to enable all the relevant features for
that platform architecture
e.g. https://git.yoctoproject.org/yocto-kernel-cache/tree/bsp/renesas-rcar?h=yocto-5.14

But for expediency to quickly get a multi platform kernel I
went with the above workaround.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>