parsec-service: configure TPM and PKCS#11 providers when available
Update /etc/parsec/config.toml on first boot:

- A TPM provider is added if /dev/tpm0 exists. This cannot be done at
build time because TPM is always selected, but it might not be available
at runtime (for example, QEMU may be used with or without swtpm). Note
that the Parsec service will fail to start if any provider fails to
initialize.

- A provider for the OP-TEE PKCS#11 interface is added if /dev/tee0
exists, and the parsec user is added to the teeclnt group so that the
parsec service may aceess the TEE device.
This could in theory be done at build time in the .bbappend file, but
is not for two reasons. First, the config fragment for TPM has to be
handled at runtime so it is more consistent to do the same for OP-TEE
PKCS#11. Second, no matter how hard I tried, I could not find how to
add the parsec user to the teeclnt group from within the .bbappend file
without breaking what is done in the main parsec-service recipe (i.e.,
adding the user to the tss group when a TPM is configured).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>