parsec-service: switch from append to += when amending to SRC_URI
Things added with :append can't be removed if someone in other layers
needs to do it. :append should only be used with additional machine
or distro specific use cases. += works here so use it.

Remove tab character from the end. bitbake runs with both
python and shell so tabs vs spaces may be a problem.

Signed-off-by: Mikko Rapeli <mikko.rapeli@linaro.org>