linux-yocto: force module signing
Use the linux-yocto way of enabling module signing and use
the force module signing variant which does not allow loading unsigned
kernel modules. Nothing currently depends on loading unsigned kernel
modules so use the production ready variant.

Upstream poky maintainers rejected enabling this feature by default
but we want it. meta-security has another way for the same but uses
a custom certificates and keys for this.

Signed-off-by: Mikko Rapeli <mikko.rapeli@linaro.org>